+++ adminer/drivers/sqlite.inc.php
@@ -728,10 +728,6 @@ if (isset($_GET["sqlite"])) {
 		);
 	}
 
-	function begin() {
-		return queries("BEGIN");
-	}
-
 	function last_id($result): string {
 		return get_val("SELECT LAST_INSERT_ROWID()");
 	}
@@ -800,6 +796,6 @@ if (isset($_GET["sqlite"])) {
 	}
 
 	function support(string $feature): bool {
-		return preg_match('~^(check|columns|database|drop_col|dump|indexes|descidx|move_col|sql|status|table|trigger|variables|view|view_trigger)$~', $feature);
+		return preg_match('~^(check|columns|database|drop_col|dump|indexes|descidx|move_col|sql|status|table|transaction_ddl|trigger|variables|view|view_trigger)$~', $feature);
 	}
 }